Rename forms/elements to forms/patterns
