+++ packages/design/src/config/edit/SubmissionConfirmationEdit.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { useFormContext } from 'react-hook-form';
 
-import { type InputElement } from '@atj/forms/src/elements/input';
+import { type InputElement } from '@atj/forms/src/patterns/input';
 import { type PatternEditComponent } from '../../FormManager/FormEdit/types';
 
 const InputElementEdit: PatternEditComponent<InputElement> = ({ element }) => {
