+++ packages/forms/src/patterns/input.ts
